Improve player movement

+++ Ported/Jump-the-Gun_Group2/Assets/Scripts/Components/Direction.cs
@@ -4,5 +4,5 @@ using Unity.Mathematics;
 [GenerateAuthoringComponent]
 public struct Direction : IComponentData
 {
-    public float2 Value;
+    public int2 Value;
 }
